requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
scikit-learn>=1.3
Pillow>=10.0
